+++ fem/StableNeo.h
@@ -10,10 +10,10 @@ class StableNeo : public StrainEne {
   std::vector<Matrix3f> getdPdx(const Matrix3f& F, const Vec3f& dF,
                                 int dim = 3);
   void CacheF(const Matrix3f& F) override;
-  void SetParam(const std::vector<double>&param)override;
-  double Mu() const { return param[0]; }
-  double Lambda() const { return param[1]; }
-  double Alpha() const { return param[2]; }
+  void SetParam(const std::vector<float>&param)override;
+  float Mu() const { return param[0]; }
+  float Lambda() const { return param[1]; }
+  float Alpha() const { return param[2]; }
 
   Matrix3f _pJpF;
